worker/requirements.txt
azure-functions
requests
azure-cosmos
azure-identity